+++ model_utils/managers.py
@@ -274,24 +274,10 @@ class SoftDeletableManagerMixin:
     """
     _queryset_class = SoftDeletableQuerySet
 
-    def __init__(self, *args, _emit_deprecation_warnings=False, **kwargs):
-        self.emit_deprecation_warnings = _emit_deprecation_warnings
-        super().__init__(*args, **kwargs)
-
     def get_queryset(self):
         """
         Return queryset limited to not removed entries.
         """
-
-        if self.emit_deprecation_warnings:
-            warning_message = (
-                "{0}.objects model manager will include soft-deleted objects in an "
-                "upcoming release; please use {0}.available_objects to continue "
-                "excluding soft-deleted objects. See "
-                "https://django-model-utils.readthedocs.io/en/stable/models.html"
-                "#softdeletablemodel for more information."
-            ).format(self.model.__class__.__name__)
-            warnings.warn(warning_message, DeprecationWarning)
 
         kwargs = {'model': self.model, 'using': self._db}
         if hasattr(self, '_hints'):