+++ CHANGES.rst
@@ -1,6 +1,11 @@
 CHANGES
 =======
 
+Unreleased
+----------
+
+- SoftDeletableModel no longer uses SoftDeletableManager as its default manager (GH-364)
+
 4.1.1 (2020-12-01)
 ------------------
 - Applied `isort` to codebase (Refs GH-#402)